chore: move available jobs to module in worker
