+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/desktop/my-jobs-table.tsx
@@ -16,21 +16,21 @@ import {
 } from '@/modules/worker/services/my-jobs-data';
 import { useMyJobsFilterStore } from '@/modules/worker/hooks/use-my-jobs-filter-store';
 import { getNetworkName } from '@/modules/smart-contracts/get-network-name';
-import { RewardAmount } from '@/modules/worker/components/jobs/reward-amount';
+import { RewardAmount } from '@/modules/worker/jobs/reward-amount';
 import { Button } from '@/shared/components/ui/button';
 import { Chip } from '@/shared/components/ui/chip';
-import { EvmAddress } from '@/modules/worker/components/jobs/evm-address';
-import { MyJobsJobTypeFilter } from '@/modules/worker/components/jobs/my-jobs/desktop/my-jobs-job-type-filter';
-import { MyJobsRewardAmountSort } from '@/modules/worker/components/jobs/my-jobs/desktop/my-jobs-reward-amount-sort';
-import { MyJobsStatusFilter } from '@/modules/worker/components/jobs/my-jobs/desktop/my-jobs-status-filter';
-import { MyJobsExpiresAtSort } from '@/modules/worker/components/jobs/my-jobs/desktop/my-jobs-expires-at-sort';
-import { MyJobsNetworkFilter } from '@/modules/worker/components/jobs/my-jobs/desktop/my-jobs-network-filter';
+import { EvmAddress } from '@/modules/worker/jobs/evm-address';
+import { MyJobsJobTypeFilter } from '@/modules/worker/jobs/my-jobs/desktop/my-jobs-job-type-filter';
+import { MyJobsRewardAmountSort } from '@/modules/worker/jobs/my-jobs/desktop/my-jobs-reward-amount-sort';
+import { MyJobsStatusFilter } from '@/modules/worker/jobs/my-jobs/desktop/my-jobs-status-filter';
+import { MyJobsExpiresAtSort } from '@/modules/worker/jobs/my-jobs/desktop/my-jobs-expires-at-sort';
+import { MyJobsNetworkFilter } from '@/modules/worker/jobs/my-jobs/desktop/my-jobs-network-filter';
 import { useColorMode } from '@/shared/contexts/color-mode';
 import { createTableDarkMode } from '@/shared/styles/create-table-dark-mode';
 import type { JobType } from '@/modules/smart-contracts/EthKVStore/config';
-import { EscrowAddressSearchForm } from '@/modules/worker/components/jobs/escrow-address-search-form';
+import { EscrowAddressSearchForm } from '@/modules/worker/jobs/escrow-address-search-form';
 import { useRefreshTasksMutation } from '@/modules/worker/services/refresh-tasks';
-import { StatusChip } from '@/modules/worker/components/jobs/status-chip';
+import { StatusChip } from '@/modules/worker/jobs/status-chip';
 import { formatDate } from '@/shared/helpers/date';
 import { MyJobsTableActions } from '../../my-jobs-table-actions';
