+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/mobile/my-jobs-drawer-mobile.tsx
@@ -7,8 +7,8 @@ import CloseIcon from '@mui/icons-material/Close';
 import type { Dispatch, SetStateAction } from 'react';
 import { HumanLogoIcon } from '@/shared/components/ui/icons';
 import { useColorMode } from '@/shared/contexts/color-mode';
-import { MyJobsRewardAmountSortMobile } from '@/modules/worker/components/jobs/my-jobs/mobile/my-jobs-reward-amount-sort-mobile';
-import { MyJobsExpiresAtSortMobile } from '@/modules/worker/components/jobs/my-jobs/mobile/my-jobs-expires-at-sort-mobile';
+import { MyJobsRewardAmountSortMobile } from '@/modules/worker/jobs/my-jobs/mobile/my-jobs-reward-amount-sort-mobile';
+import { MyJobsExpiresAtSortMobile } from '@/modules/worker/jobs/my-jobs/mobile/my-jobs-expires-at-sort-mobile';
 import { useHandleMainNavIconClick } from '@/shared/hooks/use-handle-main-nav-icon-click';
 import { MyJobsNetworkFilterMobile } from './my-jobs-network-filter-mobile';
 import { MyJobsJobTypeFilterMobile } from './my-jobs-job-type-filter-mobile';
