+++ packages/apps/human-app/frontend/src/modules/worker/jobs/reward-amount.tsx
